requirements.txt
bsdiff4
PyGithub
python-dotenv
pyelftools
west